Note: Significant test failures!
Get this branch to where it's not complaining about Reve::Methods::Urls having cattr_accessor.

Next up is to change the tests to change the variables for each new module, that is:

for test_wallet_transactions_clean:
  Reve::Methods::Char.personal_wallet_trans_url = XML_BASE + 'market_transactions.xml'

The new scoping is to have all of the URLs defined in the one urls/urls.rb file (This will probably change to Reve::Urls to drop the Methods namespace, since it's not really a method) and each dependent file can include it. Later revisions will permit dropping in modules containing URLs for each kind of method (char/corp/eve).